IRender接口调整：完善IPath接口与实现。

+++ duilib/RenderSkia/Matrix_Skia.cpp
@@ -11,7 +11,7 @@ Matrix_Skia::Matrix_Skia()
     m_matrix = std::make_unique<SkMatrix>();
 }
 
-void Matrix_Skia::Translate(int offsetX, int offsetY)
+void Matrix_Skia::Translate(float offsetX, float offsetY)
 {
     m_matrix->setTranslate((SkScalar)offsetX, (SkScalar)offsetY);
 }
