+++ duilib/RenderSkia/Matrix_Skia.h
@@ -13,7 +13,7 @@ class UILIB_API Matrix_Skia : public IMatrix
 public:
     Matrix_Skia();
 
-    virtual void Translate(int offsetX, int offsetY) override;
+    virtual void Translate(float offsetX, float offsetY) override;
     virtual void Scale(float scaleX, float scaleY) override;
     virtual void Rotate(float angle) override;
     virtual void RotateAt(float angle, const UiPoint& center) override;
